Vastly simplified structure. No reason to maintain stats if they are not used.
