+++ Source/Pawnmorphs/Esoteria/Pawnmorph.csproj
@@ -857,7 +857,7 @@
     <Compile Include="Utilities\RandUtilities.cs" />
     <Compile Include="Utilities\RWRaycast.cs" />
     <Compile Include="Utilities\SkillMod.cs" />
-    <Compile Include="Utilities\StatsCache.cs" />
+    <Compile Include="Utilities\StatsUtility.cs" />
     <Compile Include="Utilities\TimeMetrics.cs" />
     <Compile Include="Utilities\TreeUtilities.cs" />
     <Compile Include="Verbs\Tag.cs" />
